rename Global ~> Root
