+++ apps/client/src/components/common/Roles/hooks/useGuest.ts
@@ -13,7 +13,7 @@ export const useGuest = () => {
 
    const checkAuth = async () => {
       try {
-         const { request } = API['/api/user/global/auth'].get
+         const { request } = API['/api/auth-check'].get
 
          const response = await defaultAxios<TokenCheckResponse>(request)
