+++ apps/server/routes/root/services/logout.ts
